feat: added interface folder
