+++ backend/main.py
@@ -10,8 +10,8 @@ app = _fastapi.FastAPI()
 
 _database.create_database()
 
+app.include_router(_cursoRoute.router)
 app.include_router(_alunoRoute.router)
 app.include_router(_profRoute.router)
 app.include_router(_enderecoRoute.router)
 app.include_router(_telRoute.router)
-app.include_router(_cursoRoute.router)
