+++ backend/src/controller/enderecoController.py
@@ -1,7 +1,7 @@
 import sqlalchemy.orm as _orm
 from typing import List
 
-from . import IEnderecoController as _enderecoController
+from .interface import IEnderecoController as _enderecoController
 from ..dao import enderecoDAO as _enderecoDAO
 from ..schemas import enderecoSchema as _enderecoSchema
 from ..models import endereco as _enderecoModel
@@ -20,19 +20,19 @@ class EnderecoController(_enderecoController.IEnderecoController):
     def inserir(
         self, db: _orm.Session, endereco: _enderecoSchema.EnderecoCreate
     ) -> _enderecoSchema.Endereco:
-        endereco_db = self.model(**endereco.dict())
+        endereco_db = self.model(**endereco.model_dump())
         return self.endereco_dao.inserir(db, endereco_db)
 
     def atualizar(
         self, db: _orm.Session, endereco: _enderecoSchema.Endereco
     ) -> _enderecoSchema.Endereco:
-        endereco_db = self.model(**endereco.dict())
+        endereco_db = self.model(**endereco.model_dump())
         return self.endereco_dao.atualizar(db, endereco_db)
 
     def remover(
         self, db: _orm.Session, endereco: _enderecoSchema.Endereco
     ) -> _enderecoSchema.Endereco:
-        endereco_db = self.model(**endereco.dict())
+        endereco_db = self.model(**endereco.model_dump())
         return self.endereco_dao.remover(db, endereco_db)
 
     def buscar(self, db: _orm.Session, endereco_id: int) -> _enderecoSchema.Endereco:
