+++ backend/src/controller/professorController.py
@@ -1,7 +1,7 @@
 import sqlalchemy.orm as _orm
 from typing import List
 
-from . import IPessoaController as _pessoaController
+from .interface import IPessoaController as _pessoaController
 from ..dao import professorDAO as _professorDAO
 from ..schemas import professorSchema as _professorSchema
 from ..models import professor as _professorModel
@@ -15,13 +15,13 @@ class ProfessorController(_pessoaController.IPessoaController):
     def inserir(
         self, db: _orm.Session, pessoa: _professorSchema.ProfessorCreate
     ) -> _professorSchema.Professor:
-        res = _professorModel.Professor(**pessoa.dict())
+        res = _professorModel.Professor(**pessoa.model_dump())
         return self.professor_dao.inserir(db, res)
 
     def atualizar(
         self, db: _orm.Session, pessoa: _professorSchema.ProfessorCreate
     ) -> _professorSchema.Professor:
-        res = _professorModel.Professor(**pessoa.dict())
+        res = _professorModel.Professor(**pessoa.model_dump())
         return self.professor_dao.atualizar(db, res)
 
     def remover(self, db: _orm.Session, pessoa_id: int) -> _professorSchema.Professor:
