+++ backend/src/controller/telefoneController.py
@@ -1,7 +1,7 @@
 import sqlalchemy.orm as _orm
 from typing import List
 
-from . import ITelefoneController as _telefoneController
+from .interface import ITelefoneController as _telefoneController
 from ..dao import telefoneDAO as _telefoneDAO
 from ..schemas import telefoneSchema as _telefoneSchema
 from ..models import telefone as _telefoneModel
@@ -20,19 +20,19 @@ class TelefoneController(_telefoneController.ITelefoneController):
     def inserir(
         self, db: _orm.Session, telefone: _telefoneSchema.TelefoneCreate
     ) -> _telefoneSchema.Telefone:
-        telefone_db = self.model(**telefone.dict())
+        telefone_db = self.model(**telefone.model_dump())
         return self.telefone_dao.inserir(db, telefone_db)
 
     def atualizar(
         self, db: _orm.Session, telefone: _telefoneSchema.Telefone
     ) -> _telefoneSchema.Telefone:
-        telefone_db = self.model(**telefone.dict())
+        telefone_db = self.model(**telefone.model_dump())
         return self.telefone_dao.atualizar(db, telefone_db)
 
     def remover(
         self, db: _orm.Session, telefone: _telefoneSchema.Telefone
     ) -> _telefoneSchema.Telefone:
-        telefone_db = self.model(**telefone.dict())
+        telefone_db = self.model(**telefone.model_dump())
         return self.telefone_dao.remover(db, telefone_db)
 
     def buscar(self, db: _orm.Session, telefone_id: int) -> _telefoneSchema.Telefone:
